+++ api/openapi_server/controllers/files_controller.py
@@ -2,17 +2,16 @@ import os
 import tempfile
 import uuid
 import subprocess
-import sys
+from locale import getpreferredencoding
 
 from flask import current_app, send_from_directory, abort
 from werkzeug.utils import secure_filename
 from werkzeug.exceptions import BadRequest
 import connexion
-import six
 
-from openapi_server import util
 
-ALLOWED_EXTENSIONS = set(['GSvar'])
+ALLOWED_EXTENSIONS = set('GSvar')
+
 
 def count_file_path_get(filePath):  # noqa: E501
     """count_file_path_get
@@ -33,14 +32,16 @@ def count_file_path_get(filePath):  # noqa: E501
 
     :rtype: float
     """
-    absFilePath = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
-    if os.path.isfile(absFilePath):
-        command = "cat {} | grep -v '#' | wc -l".format(absFilePath)
+    abs_file_path = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
+    if os.path.isfile(abs_file_path):
+        command = "cat {} | grep -v '#' | wc -l".format(abs_file_path)
         count = subprocess.check_output(command, shell=True)
-        count = str(count, sys.stdout.encoding).strip()
+        print(getpreferredencoding())
+        count = str(count, getpreferredencoding()).strip()
         return count
     else:
         abort(404)
+
 
 def download_file_path_get(filePath, lines=None):  # noqa: E501
     """download_file_path_get
@@ -54,14 +55,15 @@ def download_file_path_get(filePath, lines=None):  # noqa: E501
 
     :rtype: file
     """
-    absFilePath = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
-    if os.path.isfile(absFilePath): # this makes sure the upload folder is not escaped
+    abs_file_path = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
+    if os.path.isfile(abs_file_path):  # this makes sure the upload folder is not escaped
         lines = connexion.request.headers['Lines'] if 'lines' in connexion.request.headers else None
 
         if lines:
             lines = lines.split('-')
             if len(lines) != 2:
-                raise BadRequest("Lines should contain two elements, not {}".format(len(lines)))
+                raise BadRequest(
+                    "Lines should contain two elements, not {}".format(len(lines)))
             elif not len(lines[0]):
                 raise BadRequest("You cannot specify a negative number here..")
             elif lines[0] > lines[1]:
@@ -70,18 +72,21 @@ def download_file_path_get(filePath, lines=None):  # noqa: E501
                 raise BadRequest("Line start cannot be smaller than one")
 
             # lines[1] > file_length is handled by sed quite elegantly)
-            tmpFile = uuid.uuid4().hex
-            tmpPath = os.path.join(tempfile.gettempdir(), tmpFile)
-            status = os.system("sed -n '{},{} p' {} >> {}".format(lines[0], lines[1], absFilePath, tmpPath))
+            tmp_file = uuid.uuid4().hex
+            tmp_path = os.path.join(tempfile.gettempdir(), tmp_file)
+            status = os.system(
+                "sed -n '{},{} p' {} >> {}".format(lines[0], lines[1], abs_file_path, tmp_path))
             if status == 0:
-                return send_from_directory(directory=tempfile.gettempdir(), filename=tmpFile)
+                return send_from_directory(directory=tempfile.gettempdir(), filename=tmp_file)
             else:
-                raise BadRequest("Command exited with status {}".format(status))
+                raise BadRequest(
+                    "Command exited with status {}".format(status))
         else:
             return send_from_directory(directory=current_app.config['UPLOAD_FOLDER'], filename=filePath)
     else:
         abort(404)
     return 'do some magic!'
+
 
 def upload_post(uploadedFile=None):  # noqa: E501
     """upload_post
@@ -93,11 +98,13 @@ def upload_post(uploadedFile=None):  # noqa: E501
 
     :rtype: None
     """
-    fileName = secure_filename(uploadedFile.filename)
-    absFilePath = os.path.join(current_app.config['UPLOAD_FOLDER'], fileName)
-    if os.path.exists(absFilePath):
+    file_name = secure_filename(uploadedFile.filename)
+    abs_file_path = os.path.join(
+        current_app.config['UPLOAD_FOLDER'], file_name)
+    if os.path.exists(abs_file_path):
         pass
     else:
-        uploadedFile.save(absFilePath) # this will raise a IOError if something goes wrong
+        # this will raise a IOError if something goes wrong
+        uploadedFile.save(abs_file_path)
 
     return "successfull"