+++ api/openapi_server/controllers/meg_sap_controller.py
@@ -1,11 +1,8 @@
 import os
 
-import connexion
-import six
 from flask import current_app, abort
 from werkzeug.exceptions import BadRequest
 
-from openapi_server import util
 
 def an_vep_file_path_get(filePath):  # noqa: E501
     """an_vep_file_path_get
@@ -17,21 +14,24 @@ def an_vep_file_path_get(filePath):  # noqa: E501
 
     :rtype: string
     """
-    absFilePath = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
-    if os.path.isfile(absFilePath):
-        absOutPath = absFilePath.replace(".vcf", "_annotated.vcf")
-        if not os.path.isfile(absOutPath):
-            megSAP = os.path.abspath(os.getenv('MEGSAP_DIR'))
-            megSAPcommand = "php {}/src/NGS/an_vep.php -in {} -out {}".format(megSAP, absFilePath, absOutPath)
-            status = os.system(megSAPcommand)
+    abs_file_path = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
+    if os.path.isfile(abs_file_path):
+        abs_out_path = abs_file_path.replace(".vcf", "_annotated.vcf")
+        if not os.path.isfile(abs_out_path):
+            meg_sap = os.path.abspath(os.getenv('MEGSAP_DIR'))
+            meg_sap_command = "php {}/src/NGS/an_vep.php -in {} -out {}".format(
+                meg_sap, abs_file_path, abs_out_path)
+            status = os.system(meg_sap_command)
             if status == 0:
-                return os.path.basename(absOutPath)
+                return os.path.basename(abs_out_path)
             else:
-                raise BadRequest("Command exited with status {}".format(status))
+                raise BadRequest(
+                    "Command exited with status {}".format(status))
         else:
-            return os.path.basename(absOutPath)
+            return os.path.basename(abs_out_path)
     else:
         abort(404)
+
 
 def vcf2gsvar_file_path_get(filePath):  # noqa: E501
     """vcf2gsvar_file_path_get
@@ -44,26 +44,30 @@ def vcf2gsvar_file_path_get(filePath):  # noqa: E501
     :rtype: string
     """
 
-    absFilePath = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
-    if os.path.isfile(absFilePath):
-        absGSvarPath = absFilePath.replace(".vcf", ".GSvar")
-        if not os.path.isfile(absGSvarPath):
-            binFolder = os.path.abspath(os.getenv('NGS_BITS_BIN', os.getcwd()))
-            command = "./VcfCheck -in {}".format(absFilePath)
-            fullCommand = "cd {} && {}".format(binFolder, command)
+    abs_file_path = os.path.join(current_app.config['UPLOAD_FOLDER'], filePath)
+    if os.path.isfile(abs_file_path):
+        abs_gsvar_path = abs_file_path.replace(".vcf", ".GSvar")
+        if not os.path.isfile(abs_gsvar_path):
+            bin_folder = os.path.abspath(
+                os.getenv('NGS_BITS_BIN', os.getcwd()))
+            command = "./VcfCheck -in {}".format(abs_file_path)
+            full_command = "cd {} && {}".format(bin_folder, command)
 
-            status = os.system(fullCommand)
+            status = os.system(full_command)
             if status != 0:
-                raise BadRequest("Command exited with status {} while checking the VCF".format(status))
+                raise BadRequest(
+                    "Command exited with status {} while checking the VCF".format(status))
 
-            megSAP = os.path.abspath(os.getenv('MEGSAP_DIR'))
-            megSAPcommand = "php {}/src/NGS/vcf2gsvar.php -in {} -out {}".format(megSAP, absFilePath, absGSvarPath)
-            status = os.system(megSAPcommand)
+            meg_sap = os.path.abspath(os.getenv('MEGSAP_DIR'))
+            meg_sap_command = "php {}/src/NGS/vcf2gsvar.php -in {} -out {}".format(
+                meg_sap, abs_file_path, abs_gsvar_path)
+            status = os.system(meg_sap_command)
             if status == 0:
-                return os.path.basename(absGSvarPath)
+                return os.path.basename(abs_gsvar_path)
             else:
-                raise BadRequest("Command failed while producing a GSvar file with status {}".format(status))
+                raise BadRequest(
+                    "Command failed while producing a GSvar file with status {}".format(status))
         else:
-            return os.path.basename(absGSvarPath)
+            return os.path.basename(abs_gsvar_path)
     else:
         abort(404)