+++ api/openapi_server/models/inline_object.py
@@ -1,9 +1,6 @@
 # coding: utf-8
 
 from __future__ import absolute_import
-from datetime import date, datetime  # noqa: F401
-
-from typing import List, Dict  # noqa: F401
 
 from openapi_server.models.base_model_ import Model
 from openapi_server import util
@@ -22,7 +19,7 @@ class InlineObject(Model):
         :type uploaded_file: file
         """
         self.openapi_types = {
-            'uploaded_file': file
+            'uploaded_file': file  # noqa: F821
         }
 
         self.attribute_map = {