+++ api/openapi_server/test/__init__.py
@@ -1,3 +1,4 @@
+import os
 import logging
 
 import connexion
@@ -11,6 +12,8 @@ class BaseTestCase(TestCase):
     def create_app(self):
         logging.getLogger('connexion.operation').setLevel('ERROR')
         app = connexion.App(__name__, specification_dir='../openapi/')
+        app.app.config['UPLOAD_FOLDER'] = os.path.abspath(
+            os.getenv('NGS_BITS_DATA', os.getcwd()))
         app.app.json_encoder = JSONEncoder
-        app.add_api('openapi.yaml', pythonic_params=True)
+        app.add_api('openapi.yaml')
         return app.app