+++ api/openapi_server/test/test_files_controller.py
@@ -3,7 +3,6 @@
 from __future__ import absolute_import
 import unittest
 
-from flask import json
 from six import BytesIO
 
 from openapi_server.test import BaseTestCase
@@ -15,13 +14,15 @@ class TestFilesController(BaseTestCase):
     def test_count_file_path_get(self):
         """Test case for count_file_path_get
 
-        
+
         """
-        headers = { 
+        headers = {
             'Accept': 'application/json',
+            'Content-Type': 'application/json',
+            'Authorization': 'Bearer abc'
         }
         response = self.client.open(
-            '/v1/count/{file_path}'.format(file_path='file_path_example'),
+            '/v1/count/{}'.format('.travis.yml'),
             method='GET',
             headers=headers)
         self.assert200(response,
@@ -30,14 +31,15 @@ class TestFilesController(BaseTestCase):
     def test_download_file_path_get(self):
         """Test case for download_file_path_get
 
-        
+
         """
-        headers = { 
+        headers = {
             'Accept': 'text/gsvar',
-            'lines': 'lines_example',
+            'Lines': '1-100',
+            'Authorization': 'Bearer abc'
         }
         response = self.client.open(
-            '/v1/download/{file_path}'.format(file_path='file_path_example'),
+            '/v1/download/{}'.format('.travis.yml'),
             method='GET',
             headers=headers)
         self.assert200(response,
@@ -47,11 +49,12 @@ class TestFilesController(BaseTestCase):
     def test_upload_post(self):
         """Test case for upload_post
 
-        
+
         """
-        headers = { 
+        headers = {
             'Accept': '*/*',
             'Content-Type': 'multipart/form-data',
+            'Authorization': 'Bearer abc'
         }
         data = dict(uploaded_file=(BytesIO(b'some file data'), 'file.txt'))
         response = self.client.open(