+++ api/openapi_server/tools/import_and_convert.py
@@ -4,52 +4,54 @@ def convert_lines_to_dict(content):
     """
 
     filters = []
-    currentFilterList = {}
-    currentFilter = []
-    currentFilterName = ""
+    current_filter_list = {}
+    current_filter = []
+    current_filter_name = ""
 
     for index, line in enumerate(content):
         line = line.rstrip()
 
         if line.startswith("#---"):
-            if len(currentFilter):
-                currentFilterList[currentFilterName] = currentFilter
-            filters += [currentFilterList]
-            currentFilterList = {}
-            if len(currentFilter):
-                currentFilter = []
+            if len(current_filter):
+                current_filter_list[current_filter_name] = current_filter
+            filters += [current_filter_list]
+            current_filter_list = {}
+            if len(current_filter):
+                current_filter = []
         elif line.startswith("#"):
             line = line.lstrip("#")
-            if len(currentFilter):
-                currentFilterList[currentFilterName] = currentFilter
-                currentFilter = []
-                currentFilterName = line
+            if len(current_filter):
+                current_filter_list[current_filter_name] = current_filter
+                current_filter = []
+                current_filter_name = line
             else:
-                currentFilterName = line
+                current_filter_name = line
         elif len(line):
             fields = line.split("\t")
-            currentFilter.append({ 'name': fields[0], 'content': fields[1:]})
+            current_filter.append({'name': fields[0], 'content': fields[1:]})
         elif index == (len(content) - 1):
-            currentFilterList[currentFilterName] = currentFilter
-            filters += [currentFilterList]
-        else: # ingore empty lines
+            current_filter_list[current_filter_name] = current_filter
+            filters += [current_filter_list]
+        else:  # ignore empty lines
             pass
 
     return filters
+
 
 def convert_dict_to_lines(dictionary):
     """
     Converts a JSON to a filter definition file
     """
     lines = []
-    lastPossibleIndex = len(dictionary) - 1
+    last_possible_index = len(dictionary) - 1
 
     for index, filterGroups in enumerate(dictionary):
-        for filter in filterGroups.keys():
-            lines.append("#{}".format(filter))
-            for property in filterGroups[filter]:
-                lines.append("{}\t".format(property['name']) + "\t".join(property['content']))
+        for filterGroup in filterGroups.keys():
+            lines.append("#{}".format(filterGroup))
+            for filterProperty in filterGroups[filterGroup]:
+                lines.append("{}\t".format(
+                    filterProperty['name']) + "\t".join(filterProperty['content']))
             lines.append("")
-        if index < lastPossibleIndex:
+        if index < last_possible_index:
             lines.append("#---")
     return "\n".join(lines)